# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Xiangdi Multimodal Agent API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8787/api/v1
paths:
  /health:
    get:
      summary: Agent service health and active provider
      responses:
        "200":
          description: Healthy
  /assets:
    post:
      summary: Upload an image or audio asset before an Agent run
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Asset accepted
  /audio/transcriptions:
    post:
      summary: Transcribe a recorded voice question
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
      responses:
        "200":
          description: Transcript
  /agent/runs:
    post:
      summary: Create one multimodal Agent run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, request_id, input, context]
              properties:
                session_id:
                  type: string
                request_id:
                  type: string
                input:
                  type: array
                  items:
                    oneOf:
                      - type: object
                        required: [type, text]
                        properties:
                          type: { const: text }
                          text: { type: string }
                      - type: object
                        required: [type, asset_id]
                        properties:
                          type: { enum: [image, audio] }
                          asset_id: { type: string }
                context:
                  type: object
      responses:
        "202":
          description: Run created
  /agent/runs/{runId}/events:
    get:
      summary: Stream Agent events through Server-Sent Events
      parameters:
        - in: path
          name: runId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: SSE stream using contracts/agent-events.schema.json
          content:
            text/event-stream:
              schema:
                type: string